Start Wrapping FFI with Idomatic API
Creates a class to represent a given webview. This should probably be
created using a builder API as the constructor is a bit of a beast
right now. Simplfied somewhat by the use of better types for some
parameters. Content (uri/html) should probably be abstracted and
provide support for automatically creating a data uri.